Lagt till extra endpoints i ItemController, ändrat återigen i Dockerfile, lagt till Items.http

+++ sql/1-create-database.sql
@@ -1,1 +1,6 @@
-create table item (id bigint not null auto_increment, name varchar(255), price float(23), primary key (id));
+create table item (
+    id bigint not null auto_increment,
+    name varchar(255),
+    price float(53),
+    primary key (id)
+);
